refactor(core): split DomModel into DomTokenBridge and DomSelectionBridge
DomModel was a thin facade over two unrelated jobs: token-to-DOM
indexing (with ref-callback registration and composition state) and
selection mechanics (boundary translation, caret placement, DOM event
wiring). This change deletes the facade, lands two real bridges, and
slims SelectionController to reactive state + orchestration.

- features/bridge/DomTokenBridge: indexing, controlFor/childrenFor
  registries, compositionStarted/Ended/isComposing, setSelecting flag,
  reconcile.
- features/selection/DomSelectionBridge: DomBoundary owner, readRaw,
  rawPositionFromBoundary, readSelectedContent, caret placement
  (#placeCollapsed, #placeExtended, #applyPreferredAddress, etc.), and
  DOM event listeners (#trackSelection, #trackUserSelecting,
  #focusEmptyEditorOnClick).
- SelectionController owns isUserSelecting, range, position,
  isAllSelected, and pushes isUserSelecting to bridge.setSelecting via
  a one-line watch in onMounted.
- Store no longer has a 'dom' field. Adapters and indexing-flavoured
  features call store.bridge.*; selection-flavoured features call
  store.selection.*.
- DomBoundary, textOffsets, isTextTokenSpan re-homed to their callers.
- features/dom/ removed.

Decisions:
- roleFor is public on DomTokenBridge.
- placeAtAddress: bridge resolves and arms the address hint, returns
  the resolved Range; controller writes range signal.
- isTextTokenSpan moved without changing top-level barrel surface
  (still dead code as far as @markput/core's exports go).

Test redistribution and doc rewrites land in follow-up commits.
