+++ packages/core/src/features/bridge/index.ts
@@ -1,1 +1,3 @@
+export {DomTokenBridge} from './DomTokenBridge'
+export type {PathElements, RegisteredRole} from './DomTokenBridge'
 export {isTextTokenSpan} from './isTextTokenSpan'
