--- packages/core/src/features/dom/DomModel.spec.ts
@@ -1,462 +1,0 @@
-import {describe, it, expect, beforeEach, vi} from 'vitest'
-
-import {watch} from '../../shared/signals'
-import {Store} from '../../store/Store'
-
-function enableStructuralStore(value: string, props: Parameters<Store['props']['set']>[0] = {}) {
-	const store = new Store()
-	store.props.set({defaultValue: value, ...props})
-	return store
-}
-
-function mountStructuralInline(value: string) {
-	const store = enableStructuralStore(value)
-	const container = document.createElement('div')
-	const textSurface = document.createElement('span')
-	container.append(textSurface)
-	document.body.append(container)
-	store.host.container(container)
-	store.host.rendered()
-	const textNode = textSurface.firstChild
-	if (!(textNode instanceof Text)) throw new Error('Structural text surface did not render a text node')
-	return {store, container, textSurface, textNode}
-}
-
-function mountStructuralInlineMark(value = 'hello @[world]') {
-	const store = enableStructuralStore(value, {Mark: () => null, options: [{markup: '@[__value__]'}]})
-	const container = document.createElement('div')
-	const before = document.createElement('span')
-	const mark = document.createElement('mark')
-	const after = document.createElement('span')
-	container.append(before, mark, after)
-	document.body.append(container)
-	store.host.container(container)
-	store.host.rendered()
-	return {store, container, before, mark, after}
-}
-
-function mountStructuralNested(value = '@[before @[nested] after]') {
-	const store = enableStructuralStore(value, {Mark: () => null, options: [{markup: '@[__slot__]'}]})
-	const container = document.createElement('div')
-	const leading = document.createElement('span')
-	const outer = document.createElement('mark')
-	const before = document.createElement('span')
-	const inner = document.createElement('mark')
-	const after = document.createElement('span')
-	const trailing = document.createElement('span')
-	outer.append(before, inner, after)
-	container.append(leading, outer, trailing)
-	document.body.append(container)
-	store.host.container(container)
-	store.host.rendered()
-	return {store, container, leading, outer, before, inner, after, trailing}
-}
-
-function mountStructuralNestedWithChildSequence(value = '@[before @[nested] after]') {
-	const store = enableStructuralStore(value, {Mark: () => null, options: [{markup: '@[__slot__]'}]})
-	const container = document.createElement('div')
-	const leading = document.createElement('span')
-	const outer = document.createElement('mark')
-	const control = document.createElement('input')
-	const host = document.createElement('span')
-	const before = document.createElement('span')
-	const inner = document.createElement('mark')
-	const after = document.createElement('span')
-	const trailing = document.createElement('span')
-	control.type = 'checkbox'
-	host.style.display = 'contents'
-	host.append(before, inner, after)
-	outer.append(control, host)
-	container.append(leading, outer, trailing)
-	document.body.append(container)
-	store.host.container(container)
-	store.dom.childrenFor([1])(host)
-	store.host.rendered()
-	return {store, container, leading, outer, control, host, before, inner, after, trailing}
-}
-
-function mountStructuralNestedWithDuplicateChildSequences(value = '@[before @[nested] after]') {
-	const store = enableStructuralStore(value, {Mark: () => null, options: [{markup: '@[__slot__]'}]})
-	const container = document.createElement('div')
-	const leading = document.createElement('span')
-	const outer = document.createElement('mark')
-	const hostA = document.createElement('span')
-	const hostB = document.createElement('span')
-	const trailing = document.createElement('span')
-	outer.append(hostA, hostB)
-	container.append(leading, outer, trailing)
-	document.body.append(container)
-	store.host.container(container)
-	store.dom.childrenFor([1])(hostA)
-	store.dom.childrenFor([1])(hostB)
-	return {store, container, outer, hostA, hostB}
-}
-
-function mountStructuralNestedWithOutsideChildSequence(value = '@[before @[nested] after]') {
-	const store = enableStructuralStore(value, {Mark: () => null, options: [{markup: '@[__slot__]'}]})
-	const container = document.createElement('div')
-	const leading = document.createElement('span')
-	const outer = document.createElement('mark')
-	const outsideHost = document.createElement('span')
-	const trailing = document.createElement('span')
-	leading.append(outsideHost)
-	container.append(leading, outer, trailing)
-	document.body.append(container)
-	store.host.container(container)
-	store.dom.childrenFor([1])(outsideHost)
-	return {store, container, outer, outsideHost}
-}
-
-function mountStructuralBlockWithControl(value: string) {
-	const store = enableStructuralStore(value, {layout: 'block'})
-	const container = document.createElement('div')
-	const row = document.createElement('div')
-	const control = document.createElement('button')
-	const textSurface = document.createElement('span')
-	control.textContent = 'x'
-	row.append(control, textSurface)
-	container.append(row)
-	document.body.append(container)
-	store.host.container(container)
-	store.dom.controlFor([0])(control)
-	store.host.rendered()
-	const textNode = textSurface.firstChild
-	const controlText = control.firstChild
-	if (!(textNode instanceof Text)) throw new Error('Structural block text surface did not render a text node')
-	if (!(controlText instanceof Text)) throw new Error('Structural control did not render a text node')
-	return {store, container, row, control, controlText, textSurface, textNode}
-}
-
-function mountStructuralBlockWithControls(value: string) {
-	const store = enableStructuralStore(value, {layout: 'block'})
-	const container = document.createElement('div')
-	const row = document.createElement('div')
-	const beforeControl = document.createElement('button')
-	const afterControl = document.createElement('button')
-	const textSurface = document.createElement('span')
-	beforeControl.textContent = 'before'
-	afterControl.textContent = 'after'
-	row.append(beforeControl, textSurface, afterControl)
-	container.append(row)
-	document.body.append(container)
-	store.host.container(container)
-	store.dom.controlFor([0])(beforeControl)
-	store.dom.controlFor([0])(afterControl)
-	store.host.rendered()
-	const textNode = textSurface.firstChild
-	if (!(textNode instanceof Text)) throw new Error('Structural block text surface did not render a text node')
-	return {store, container, row, beforeControl, afterControl, textSurface, textNode}
-}
-
-describe('DomModel structural indexing', () => {
-	let store: Store
-
-	beforeEach(() => {
-		vi.clearAllMocks()
-		store = new Store()
-		store.props.set({Mark: () => null, options: [{markup: '@[__value__]'}]})
-		store.value.current('hello @[world]')
-	})
-
-	it('publishes one dom index per rendered commit', () => {
-		const {store, textSurface, container} = mountStructuralInline('hello')
-
-		expect(store.dom.isIndexed()).toBe(true)
-		expect(store.dom.locateNode(textSurface)).toMatchObject({ok: true})
-		container.remove()
-	})
-
-	it('maps inline token roots by rendered token order', () => {
-		const {store, before, mark, after, container} = mountStructuralInlineMark()
-
-		expect(store.dom.locateNode(before)).toMatchObject({ok: true, value: {tokenElement: before}})
-		expect(store.dom.locateNode(mark)).toMatchObject({ok: true, value: {tokenElement: mark}})
-		expect(store.dom.locateNode(after)).toMatchObject({ok: true, value: {tokenElement: after}})
-		container.remove()
-	})
-
-	it('treats text token roots as editable text surfaces', () => {
-		const {store, textSurface, container} = mountStructuralInline('hello')
-
-		expect(store.dom.locateNode(textSurface)).toMatchObject({
-			ok: true,
-			value: {tokenElement: textSurface, textElement: textSurface},
-		})
-		expect(textSurface.textContent).toBe('hello')
-		expect(textSurface.contentEditable).toBe('true')
-		container.remove()
-	})
-
-	it('maps nested children without slot-root wrappers', () => {
-		const {store, outer, before, inner, after, container} = mountStructuralNested()
-
-		expect(store.dom.locateNode(outer)).toMatchObject({ok: true, value: {tokenElement: outer}})
-		expect(store.dom.locateNode(before)).toMatchObject({ok: true, value: {tokenElement: before}})
-		expect(store.dom.locateNode(inner)).toMatchObject({ok: true, value: {tokenElement: inner}})
-		expect(store.dom.locateNode(after)).toMatchObject({ok: true, value: {tokenElement: after}})
-		container.remove()
-	})
-
-	it('indexes nested children from a registered child sequence host', () => {
-		const {store, container, outer, control, host, before, inner, after} = mountStructuralNestedWithChildSequence()
-
-		expect(store.dom.locateNode(outer)).toMatchObject({ok: true, value: {tokenElement: outer}})
-		expect(store.dom.locateNode(host)).toMatchObject({ok: true, value: {tokenElement: outer}})
-		expect(store.dom.locateNode(control)).toMatchObject({ok: true, value: {tokenElement: outer}})
-		expect(store.dom.locateNode(before)).toMatchObject({ok: true, value: {tokenElement: before}})
-		expect(store.dom.locateNode(inner)).toMatchObject({ok: true, value: {tokenElement: inner}})
-		expect(store.dom.locateNode(after)).toMatchObject({ok: true, value: {tokenElement: after}})
-		expect(before.textContent).toBe('before ')
-		expect(after.textContent).toBe(' after')
-		container.remove()
-	})
-
-	it('maps registered child sequence host boundaries to nested child positions', () => {
-		const {store, container, host} = mountStructuralNestedWithChildSequence()
-		const tokenIndex = store.tokens.index()
-		const beforeToken = tokenIndex.resolve([1, 0])
-		const innerToken = tokenIndex.resolve([1, 1])
-		const afterToken = tokenIndex.resolve([1, 2])
-
-		expect(beforeToken?.position.end).toBe(9)
-		expect(innerToken?.position.start).toBe(9)
-		expect(innerToken?.position.end).toBe(18)
-		expect(afterToken?.position.start).toBe(18)
-		expect(store.dom.rawPositionFromBoundary(host, 1, 'before')).toEqual({
-			ok: true,
-			value: beforeToken?.position.end,
-		})
-		expect(store.dom.rawPositionFromBoundary(host, 1, 'after')).toEqual({
-			ok: true,
-			value: innerToken?.position.start,
-		})
-		expect(store.dom.rawPositionFromBoundary(host, 2, 'before')).toEqual({
-			ok: true,
-			value: innerToken?.position.end,
-		})
-		expect(store.dom.rawPositionFromBoundary(host, 2, 'after')).toEqual({
-			ok: true,
-			value: afterToken?.position.start,
-		})
-		container.remove()
-	})
-
-	it('completes indexing when duplicate child sequence hosts are registered', () => {
-		const {store, container, outer} = mountStructuralNestedWithDuplicateChildSequences()
-
-		store.host.rendered()
-
-		expect(store.dom.isIndexed()).toBe(true)
-		expect(store.dom.locateNode(outer)).toMatchObject({ok: true})
-		container.remove()
-	})
-
-	it('completes indexing when child sequence host is outside owner mark root', () => {
-		const {store, container, outer} = mountStructuralNestedWithOutsideChildSequence()
-
-		store.host.rendered()
-
-		expect(store.dom.isIndexed()).toBe(true)
-		expect(store.dom.locateNode(outer)).toMatchObject({ok: true})
-		container.remove()
-	})
-
-	it('returns control for registered controls', () => {
-		const {store, control, container} = mountStructuralBlockWithControl('hello')
-
-		expect(store.dom.locateNode(control)).toEqual({ok: false, reason: 'control'})
-		container.remove()
-	})
-
-	it('excludes multiple controls owned by the same token path from block token indexing', () => {
-		const {store, beforeControl, afterControl, textSurface, container} = mountStructuralBlockWithControls('hello')
-
-		expect(store.dom.locateNode(beforeControl)).toEqual({ok: false, reason: 'control'})
-		expect(store.dom.locateNode(afterControl)).toEqual({ok: false, reason: 'control'})
-		expect(store.dom.locateNode(textSurface)).toMatchObject({
-			ok: true,
-			value: {tokenElement: textSurface, textElement: textSurface},
-		})
-		container.remove()
-	})
-
-	it('completes indexing when a nested mark renders no child elements', () => {
-		const store = enableStructuralStore('@[before @[nested] after]', {
-			Mark: () => null,
-			options: [{markup: '@[__slot__]'}],
-		})
-		const container = document.createElement('div')
-		const leading = document.createElement('span')
-		const outer = document.createElement('mark')
-		const trailing = document.createElement('span')
-		container.append(leading, outer, trailing)
-		document.body.append(container)
-		store.host.container(container)
-		store.host.rendered()
-
-		expect(store.dom.isIndexed()).toBe(true)
-		expect(store.dom.locateNode(outer)).toMatchObject({ok: true})
-		container.remove()
-	})
-
-	it('clamps OOB caret range and places at maxPos', () => {
-		const {store, container} = mountStructuralInline('hello')
-
-		store.selection.range({start: 999, end: 999})
-		store.host.rendered()
-
-		expect(store.selection.range()).toEqual({start: 5, end: 5})
-		container.remove()
-	})
-
-	it('clamps OOB selection range', () => {
-		const {store, container} = mountStructuralInline('hello')
-
-		store.selection.range({start: 999, end: 1000})
-		store.host.rendered()
-
-		expect(store.selection.range()).toEqual({start: 5, end: 5})
-		container.remove()
-	})
-
-	it('applies caret.selection to DOM after render', () => {
-		const {store, container, textNode} = mountStructuralInline('hello')
-
-		store.selection.range({start: 3, end: 3})
-		store.host.rendered()
-
-		const sel = window.getSelection()
-		expect(sel?.focusNode).toBe(textNode)
-		expect(sel?.focusOffset).toBe(3)
-		container.remove()
-	})
-
-	it('clamps OOB range and places caret at clamped position', () => {
-		const {store, container} = mountStructuralInline('hello') // length 5
-		store.selection.range({start: 999, end: 999})
-		store.host.rendered()
-
-		// clamped to maxPos (5); structural equality prevents re-fire
-		expect(store.selection.range()).toEqual({start: 5, end: 5})
-		container.remove()
-	})
-
-	it('skips apply when isUserSelecting', () => {
-		const {store, container} = mountStructuralInline('hello')
-		store.selection.range({start: 2, end: 2})
-		store.dom.isUserSelecting(true)
-		store.host.rendered()
-
-		expect(store.selection.range()).toEqual({start: 2, end: 2})
-		container.remove()
-	})
-
-	describe('raw boundary mapping', () => {
-		it('maps text-surface boundaries to raw UTF-16 positions', () => {
-			const {store, container, textNode} = mountStructuralInline('hello')
-
-			expect(store.dom.rawPositionFromBoundary(textNode, 2)).toEqual({ok: true, value: 2})
-			container.remove()
-		})
-
-		it('rejects boundaries that split surrogate pairs', () => {
-			const {store, container, textNode} = mountStructuralInline('a😀b')
-
-			expect(store.dom.rawPositionFromBoundary(textNode, 2)).toEqual({ok: false, reason: 'invalidBoundary'})
-			container.remove()
-		})
-
-		it('maps token shell boundaries by affinity', () => {
-			const {store, container, textSurface} = mountStructuralInline('hello')
-
-			expect(store.dom.rawPositionFromBoundary(textSurface, 0, 'before')).toEqual({ok: true, value: 0})
-			expect(store.dom.rawPositionFromBoundary(textSurface, 1, 'after')).toEqual({ok: true, value: 5})
-			container.remove()
-		})
-
-		it('rejects editable boundaries inside mark presentation descendants', () => {
-			const {store, container, mark} = mountStructuralInlineMark('@[world]')
-			const descendant = document.createElement('span')
-			descendant.contentEditable = 'true'
-			descendant.textContent = 'inner'
-			mark.append(descendant)
-			const descendantText = descendant.firstChild
-			if (!(descendantText instanceof Text)) throw new Error('Mark descendant did not render a text node')
-			store.host.rendered()
-
-			expect(store.dom.rawPositionFromBoundary(descendantText, 0, 'after')).toEqual({
-				ok: false,
-				reason: 'invalidBoundary',
-			})
-			container.remove()
-		})
-
-		it('returns mixedBoundary for selections crossing controls', () => {
-			const {store, container, textNode, controlText} = mountStructuralBlockWithControl('hello')
-			const selection = window.getSelection()!
-			const range = document.createRange()
-			range.setStart(textNode, 0)
-			range.setEnd(controlText, 1)
-			selection.removeAllRanges()
-			selection.addRange(range)
-
-			expect(store.dom.readRawSelection()).toEqual({ok: false, reason: 'mixedBoundary'})
-			container.remove()
-		})
-	})
-
-	describe('indexed event and reconcile opts', () => {
-		it('indexed event fires after commitRendered', () => {
-			const store = new Store()
-			const container = document.createElement('div')
-			document.body.appendChild(container)
-			store.props.set({defaultValue: 'hi'})
-			store.host.container(container)
-
-			const fired = vi.fn()
-			watch(store.dom.indexed, fired)
-			store.host.rendered()
-			expect(fired).toHaveBeenCalledTimes(1)
-			container.remove()
-		})
-
-		it('reconcile respects isUserSelecting signal', () => {
-			const store = new Store()
-			const container = document.createElement('div')
-			const span = document.createElement('span')
-			container.appendChild(span)
-			document.body.appendChild(container)
-			store.props.set({defaultValue: 'hello'})
-			store.host.container(container)
-			store.host.rendered()
-
-			store.dom.isUserSelecting(true)
-			expect(span.contentEditable).toBe('false')
-
-			store.dom.isUserSelecting(false)
-			expect(span.contentEditable).toBe('true')
-
-			container.remove()
-		})
-	})
-
-	describe('empty-editor click handler', () => {
-		it('focuses first child on click when editor is empty', () => {
-			const store = new Store()
-			const container = document.createElement('div')
-			const span = document.createElement('span')
-			span.contentEditable = 'true'
-			container.appendChild(span)
-			document.body.appendChild(container)
-
-			store.props.set({defaultValue: ''})
-			store.host.container(container)
-			store.host.rendered()
-
-			const focusSpy = vi.spyOn(span, 'focus')
-			container.dispatchEvent(new MouseEvent('click', {bubbles: true}))
-			expect(focusSpy).toHaveBeenCalledTimes(1)
-			container.remove()
-		})
-	})
-})
