--- packages/core/src/features/dom/index.ts
@@ -1,2 +1,0 @@
-export {DomModel} from './DomModel'
-export {isTextTokenSpan} from '../bridge/isTextTokenSpan'
