--- packages/core/src/features/dom/isTextTokenSpan.ts
@@ -1,6 +1,0 @@
-export function isTextTokenSpan(el: HTMLElement): boolean {
-	return (
-		el.tagName === 'SPAN' &&
-		(el.attributes.length === 0 || (el.attributes.length === 1 && el.hasAttribute('contenteditable')))
-	)
-}
