+++ packages/core/src/features/keyboard/input.ts
@@ -3,7 +3,7 @@ import type {Range} from '../../shared/editorContracts'
 import {listen} from '../../shared/signals/index.js'
 import type {Store} from '../../store/Store'
 
-type KbCtx = Pick<Store, 'dom' | 'value' | 'selection' | 'edit' | 'props' | 'tokens'>
+type KbCtx = Pick<Store, 'bridge' | 'value' | 'selection' | 'edit' | 'props' | 'tokens'>
 import {captureMarkupPaste, consumeMarkupPaste} from '../clipboard'
 import type {Token} from '../parsing'
 import {rawRangeFromInputEvent} from './inputRange'
@@ -22,15 +22,15 @@ export function enableInput(store: KbCtx, container: HTMLElement): void {
 	})
 
 	listen(container, 'compositionstart', () => {
-		const selection = store.dom.readRawSelection()
+		const selection = store.selection.readRaw()
 		compositionRange = selection.ok ? selection.value.range : undefined
-		store.dom.compositionStarted()
+		store.bridge.compositionStarted()
 	})
 
 	listen(container, 'compositionend', e => {
 		const range = compositionRange
 		compositionRange = undefined
-		store.dom.compositionEnded()
+		store.bridge.compositionEnded()
 		if (store.props.layout.isBlock()) return
 		if (!range) return
 		const data = e.data
@@ -61,7 +61,7 @@ function handleDeleteKey(store: KbCtx, event: KeyboardEvent): void {
 		return
 	}
 
-	const raw = store.dom.readRawSelection()
+	const raw = store.selection.readRaw()
 	if (!raw.ok) return
 
 	const inputType = event.key === KEYBOARD.BACKSPACE ? 'deleteContentBackward' : 'deleteContentForward'
