+++ packages/core/src/features/selection/textOffsets.ts
