+++ packages/react/markput/src/components/DropIndicator.tsx
@@ -7,9 +7,9 @@ import styles from '@markput/core/styles.module.css'
 
 export const DropIndicator = memo(({token, position}: {token: TokenType; position: 'before' | 'after'}) => {
 	const dropPosition = useMarkput(s => s.blocks.get(token).state.dropPosition)
-	const {dom, index} = useMarkput(s => ({dom: s.dom, index: s.tokens.index}))
+	const {bridge, index} = useMarkput(s => ({bridge: s.bridge, index: s.tokens.index}))
 	const path = index.pathFor(token)
-	const controlRef = useMemo(() => (path ? dom.controlFor(path) : undefined), [dom, path])
+	const controlRef = useMemo(() => (path ? bridge.controlFor(path) : undefined), [bridge, path])
 
 	if (dropPosition !== position) return null
