+++ packages/vue/markput/src/components/DragHandle.vue
@@ -25,7 +25,7 @@ const getPanelControlRef = () => {
 	if (panelControlRef) return panelControlRef
 	const path = index.value.pathFor(props.token)
 	if (!path) return undefined
-	panelControlRef = store.dom.controlFor(path)
+	panelControlRef = store.bridge.controlFor(path)
 	return panelControlRef
 }
